Use create or update func from controller-runtime
