+++ go.mod
@@ -3,6 +3,7 @@ module fossil.or.id/ghost-operator
 go 1.13
 
 require (
+	github.com/go-logr/logr v0.1.0
 	github.com/go-openapi/spec v0.19.2
 	github.com/operator-framework/operator-sdk v0.13.0
 	github.com/spf13/pflag v1.0.5
